refactor: move cli specific files
