+++ src/incus-app-container-files/cli/commands/create-app-container/cidr.ts
@@ -1,4 +1,4 @@
-import { Address, Cidr, createAddress } from "./deps.ts";
+import { Address, Cidr, createAddress } from "../../../deps.ts";
 
 export function firstIp(cidr: Cidr): Address {
   const ip: undefined | string = cidr.toArray({ from: 1, limit: 1 }).at(0);
