+++ src/incus-app-container-files/cli/commands/create-app-container/options.ts
@@ -1,10 +1,10 @@
-import { AbsolutePath } from "../../../absolute-path.ts";
-import { firstIp } from "../../../cidr.ts";
+import { AbsolutePath } from "../../absolute-path.ts";
+import { firstIp } from "./cidr.ts";
 import { Address, Cidr, createAddress } from "../../../deps.ts";
-import { getNextIdmapBaseFor, IDMAP_BASE_SIZE } from "../../../idmap.ts";
+import { getNextIdmapBaseFor, IDMAP_BASE_SIZE } from "./idmap.ts";
 import { MultiArgument } from "../../../multi-argument.ts";
-import { Size } from "../../../size.ts";
-import { resolveSshKeys, SshKey, SshKeyRaw } from "../../../ssh-key.ts";
+import { Size } from "./size.ts";
+import { resolveSshKeys, SshKey, SshKeyRaw } from "./ssh-key.ts";
 
 export type CreateAppContainerInputOptions<AppsDir extends AbsolutePath> = {
   ip: string;
