+++ src/incus-app-container-files/cli/commands/create-app-container/ssh-key.ts
@@ -1,8 +1,11 @@
 import { isString } from "https://deno.land/x/fns@1.1.0/string/is-string.ts";
-import { s } from "./deps.ts";
-import { MultiArgument, resolveMultiArgument } from "./multi-argument.ts";
-import { getPasswdRowOfLocalUser } from "./passwd-row.ts";
-import { readTextFiles } from "./read-text-files.ts";
+import { s } from "../../../deps.ts";
+import {
+  MultiArgument,
+  resolveMultiArgument,
+} from "../../../multi-argument.ts";
+import { getPasswdRowOfLocalUser } from "../../../passwd-row.ts";
+import { readTextFiles } from "../../../read-text-files.ts";
 
 export type SshKeyUsernamePrefix = "gh" | "local";
 export type PrefixedUsername<Prefix extends SshKeyUsernamePrefix> =
