+++ test/ssh-key.test.ts
@@ -1,5 +1,5 @@
 import { describe, it } from "https://deno.land/std@0.220.1/testing/bdd.ts";
-import { resolveSshKey } from "../src/incus-app-container-files/ssh-key.ts";
+import { resolveSshKey } from "../src/incus-app-container-files/cli/commands/create-app-container/ssh-key.ts";
 
 describe("ssh-key", () => {
   describe("resolveSshKey", () => {
